Apparently, this Quaternion code is expected behavior, as described in http://stackoverflow.com/questions/8225619/quaternion-regular-rotation-gives-stranges-values-when-convereted-to-axis-angl
The getAngle() function should not be used the way I did, because the axis might be different than the one supplied when defining the Quaternion. I've reverted the changes to the Quaternion class and updated the TestQuat sample to reflect this.

However, the test app shows awful rounding errors when using the Quat. Would it be a wise idea to let the Quaternion class only use doubles internally?