rebuild cairo block static libs i386 and x86_64 fat for os x
The previous archives seem to indicate the versions of the libraries being used are now very well aged and would be: libpng-1.4.3, pixman-0.19.2 and cairo-1.10.0. I can't speak to the versions and needs, so rather than potentially introducing an issue by updating them, I just rebuilt the aged versions with both i386 and x86_64 slices and 10.7 as the deployment target.