update Xcode projects for built-in block samples
AV Foundation and Core Media are of course needed in these samples as well and they all deserve the settings validation and cleanup that the Cinder project samples get.

The TUIO project was in an undead state, regardless to the AV Foundation, I think because the deployment target was set to 10.6, but it definitely wasn't in good shape.