add min/max x/y rect methods
Calculating a dynamic layout with a series of Rect is a little complex in that a Rect is represented by two points, not a point of origin and a size. The two points have a canonical order but it is not a strict requirement, so further checks are required.

The additional accessors added here don't do something that couldn't be done in user-code, but simplify and make more explicit the calculations.