Added the old 'setDoNotDispose()' method back to gl::TextureBase.
It allows you to keep the generated texture on destruction of the gl::Texture itself. This is needed if you require a custom deleter for your gl::TextureRef. The only way to create one using existing Cinder code, is to call a specific gl::Texture::create() method with an existing texture target and ID. The easiest way to create such a texture is to use one of the other create() methods, then 'transfer ownership' to a newly created TextureRef. The first texture should then not dispose of its internal buffers, hence the need for this function.

It was the cleanest solution to my particular problem that I could find :)