replace manual while loop with a single selector
The while loop would repeatedly close the first window, as close would result in the window being removed from the array. One cannot simply replace this loop with fast enumeration as the mutating the enumerated array would throw an exception, but thankfully -[NSArray makeObjectsPerformSelector:] does the whole job for us.