avoid direct access to display list member var
Fixes #1074

The system displays are lazy fetched on OS X and iOS and when using the CinderViewMac it seems that the initialization does not otherwise occur on our behalf, causing a crash in the AppCocoaViewTest project.

Tested on OS X 10.11 dev seed 7 / Xcode 7 beta 5 - the iOS change is untested but using the accessor instead of the member variable seems like a good idea anyways.