Fixed write position bug in BufferRecorderNode::process()
Now mWritePos gets updated in process() through
std::atomic::compare_exchange_strong().
The update only happens if mWritePos didn't change since the beginning
of process().
if it did change, then it means mWritePos has been reset to 0 by another
thread calling start() and
therefore it must not be updated ( by adding numWriteFrames to it ) or
the new recording will start with an offset of numWriteFrames from the
beginning of the buffer.