Start and stop animation only when the app active state changes
Both `applicationDidEnterBackground` and `applicationWillEnterForeground` are called when the app is not in an active state. This change limits start/stopAnimation to being called when the app is visible to the user. This eliminates some glitches that occur when resuming a background app, which were due to the update/draw methods being called when the app was not visible.