Replace std::result_of with decltype
std::result_of is deprecated in C++17 and removed in C++20.
This could be replaced by a wrapper type trait that dispatches to
std::invoke_result_t for C++17 and later or to std::result_of for
earlier versions but since it is only used in one place for nullary
functors, decltype will be simpler.